Moved macro replacers to to macros package
