+++ endpoints/openrtb2/auction_benchmark_test.go
@@ -17,7 +17,7 @@ import (
 	"github.com/prebid/prebid-server/experiment/adscert"
 	"github.com/prebid/prebid-server/gdpr"
 	"github.com/prebid/prebid-server/hooks"
-	"github.com/prebid/prebid-server/macros/replacer"
+	"github.com/prebid/prebid-server/macros"
 	metricsConfig "github.com/prebid/prebid-server/metrics/config"
 	"github.com/prebid/prebid-server/openrtb_ext"
 	"github.com/prebid/prebid-server/stored_requests/backends/empty_fetcher"
@@ -99,7 +99,7 @@ func BenchmarkOpenrtbEndpoint(b *testing.B) {
 		currency.NewRateConverter(&http.Client{}, "", time.Duration(0)),
 		empty_fetcher.EmptyFetcher{},
 		&adscert.NilSigner{},
-		replacer.NewReplacer(),
+		macros.NewReplacer(),
 	)
 
 	endpoint, _ := NewEndpoint(
